requirements.txt
flask
openai
algoliasearch
python-dotenv
gunicorn